+++ e2e/e2etests/e2etests.go
@@ -71,7 +71,7 @@ const (
 	TestSolanaWithdrawName                                = "solana_withdraw"
 	TestSolanaWithdrawRevertExecutableReceiverName        = "solana_withdraw_revert_executable_receiver"
 	TestSolanaWithdrawAndCallName                         = "solana_withdraw_and_call"
-	TestSolanaWithdrawAndCallALTName                      = "solana_withdraw_and_call_alt"
+	TestSolanaWithdrawAndCallAddressLookupTableName       = "solana_withdraw_and_call_alt"
 	TestSolanaWithdrawAndCallInvalidTxSizeName            = "solana_withdraw_and_call_invalid_tx_size"
 	TestSolanaWithdrawAndCallInvalidMsgEncodingName       = "solana_withdraw_and_call_invalid_msg_encoding"
 	TestZEVMToSolanaCallName                              = "zevm_to_solana_call"
@@ -92,7 +92,7 @@ const (
 	TestSPLDepositAndCallRevertWithCallThatRevertsName    = "spl_deposit_and_call_revert_with_call_that_reverts"
 	TestSPLWithdrawName                                   = "spl_withdraw"
 	TestSPLWithdrawAndCallName                            = "spl_withdraw_and_call"
-	TestSPLWithdrawAndCallALTName                         = "spl_withdraw_and_call_alt"
+	TestSPLWithdrawAndCallAddressLookupTableName          = "spl_withdraw_and_call_alt"
 	TestSPLWithdrawAndCallRevertName                      = "spl_withdraw_and_call_revert"
 	TestSPLWithdrawAndCreateReceiverAtaName               = "spl_withdraw_and_create_receiver_ata"
 
@@ -744,14 +744,14 @@ var AllE2ETests = []runner.E2ETest{
 		runner.WithMinimumVersion("v29.0.0"),
 	),
 	runner.NewE2ETest(
-		TestSolanaWithdrawAndCallALTName,
-		"withdraw SOL from ZEVM and call solana program using ALT",
+		TestSolanaWithdrawAndCallAddressLookupTableName,
+		"withdraw SOL from ZEVM and call solana program using Address Lookup Table",
 		[]runner.ArgDefinition{
 			{Description: "amount in lamport", DefaultValue: "1000000"},
-			{Description: "ALT address", DefaultValue: ""},
-			{Description: "indexes of ALT accounts that are mutable", DefaultValue: ""},
+			{Description: "Address Lookup Table address", DefaultValue: ""},
+			{Description: "indexes of Address Lookup Table accounts that are mutable", DefaultValue: ""},
 		},
-		TestSolanaWithdrawAndCallALT,
+		TestSolanaWithdrawAndCallAddressLookupTable,
 		runner.WithMinimumVersion("v29.0.0"),
 	),
 	runner.NewE2ETest(
@@ -804,14 +804,14 @@ var AllE2ETests = []runner.E2ETest{
 		runner.WithMinimumVersion("v29.0.0"),
 	),
 	runner.NewE2ETest(
-		TestSPLWithdrawAndCallALTName,
-		"withdraw SPL from ZEVM and call solana program using ALT",
+		TestSPLWithdrawAndCallAddressLookupTableName,
+		"withdraw SPL from ZEVM and call solana program using Address Lookup Table",
 		[]runner.ArgDefinition{
 			{Description: "amount in spl", DefaultValue: "1000000"},
-			{Description: "ALT address", DefaultValue: ""},
-			{Description: "indexes of ALT accounts that are mutable", DefaultValue: ""},
+			{Description: "Address Lookup Table address", DefaultValue: ""},
+			{Description: "indexes of Address Lookup Table accounts that are mutable", DefaultValue: ""},
 		},
-		TestSPLWithdrawAndCallALT,
+		TestSPLWithdrawAndCallAddressLookupTable,
 		runner.WithMinimumVersion("v29.0.0"),
 	),
 	runner.NewE2ETest(