+++ e2e/runner/solana.go
@@ -29,6 +29,87 @@ const solanaNodeSyncTolerance = 30 * time.Second
 var ConnectedProgramID = solana.MustPublicKeyFromBase58("4xEw862A2SEwMjofPkUyd4NEekmVJKJsdHkK3UkAtDrc")
 var ConnectedSPLProgramID = solana.MustPublicKeyFromBase58("8iUjRRhUCn8BjrvsWPfj8mguTe9L81ES4oAUApiF8JFC")
 
+// prefundedRandomWalletAddresses contains the base58-encoded addresses of prefunded random wallets
+// used for testing big number of accounts in payload. These are randomly generated and funded in start-solana.sh script
+var prefundedRandomWalletAddresses = []string{
+	"4C2kkMnqXMfPJ8PPK5v6TCg42k5z16f2kzqVocXoSDcq",
+	"3tqcVCVz5Jztwnku1H9zpvjaWshSpHakMuX4xUJQuuuA",
+	"7duGsuv6nB3yr15EuWuHEDD7rWovpAnjuveXJ5ySZuFV",
+	"8vjuCrCKVfnBGinWjc33zLRnG8iy53wj3YWHhKqvTE7o",
+	"bzkoxG5YMeWxKfNcjzbEHb3XaGTY4NLKfejjDmxVhhY",
+	"GUjKWPmnXNwPLR6kcrkLSBARmQdYnySPpxVNEUGFLs72",
+	"5oqdTyA78hpeP8RTwRBmoxCvp1V7DFicKj7T2DvtDDQM",
+	"C481t79gpWbsWwPD9eJZTAo5TSaTBet8icEkiPhwKLDx",
+	"EJvNNovWkfQYrmyMncqVvHNde2QJpSA4EJk355vyQWph",
+	"EkUpd7HFbSYPJEDbXZeDsCG19Hj5vbKTUt4rzpPYKsTM",
+	"7c7TqqdbKRWDNLVAxNa481F355AAij1fdSRttzUNVNeD",
+	"FuefjNTywey57U2zW6SmBWaJsCx7E84jmWUCbk52sBHR",
+	"fczqc5N5arnKbvMj1kgg9P1FpYPQBmJsyTXcrmK9bbp",
+	"GyH4mpobR6g2npNo5vRNcs2Cv8CxDKbK82p34kNCB4p2",
+	"9XYF8U1srAkUETkywtFrZsApipENiMU3C8Rnz5aPib94",
+	"ABnm3PMB4onvCFriWg7eBcNSmiye9iq9rRdeBViJqWif",
+	"AEydLk3RXZv67wry7EMZmS1uHLYdH8ia6xdsYri4hyB2",
+	"ArDNFdmDzrRP13UTJ4nyP11NfjV6aiQrxNwFUnwm3h8N",
+	"FzAt9aPKFUy1D2Qq8u7myYW8HFqKLzbQS2paaWz8iAmZ",
+	"9RybduN4CJHaZXvUiHoZ7KsHS9dgv1NCSAdJZaRJDW5U",
+	"C54jMgtk2umaJYoPD8aF3hmuH8XkAz2xA4sxr2ZtJABV",
+	"8kRqLbezvj4apyaK6fanurQJjhDQwn6wnW1Yr9H96gsT",
+	"2fddFSJoGJ2YuAZWxEvK9pRXXWLSKJ44rfJNVZg5WHBn",
+	"GW9oi4yqAUFBcUNUHqz56FRdfHU6Md1t9o7i2svL1XcG",
+	"Gre1nqrE1KyBbBH2Xb4qVQWFFdyXetLR5rruJpJkNhkV",
+	"BUrJRTsFVqnuLeq4Vfs7NrrU6EsL957n4tBZGoRRxkza",
+	"3tE2kKyPfuwC5rZUpJ9NaKMBf5og7G3rVTzE9CNMm9Q9",
+	"CFEQ79VSAupXWmdzvmjNtef5BybDVYXBKeh8frNHZiYe",
+	"bTnsajQuybXV6Wf9V7a8wQrwaWZ1WskkmUUChimkWmc",
+	"4dpxhhomWY3A9ey9g7EfxRQKfXnikfM4tRrFAgQf8Y8n",
+	"aQJPrcj4LNNHh9UK41sfcACFspaFR7wgcUTgmSKRXiB",
+	"5ndhaFZ48eKyU7f66vq7WSbjZRh9WhpnbgdMwDRrvgj4",
+	"G7m7dSWH5tb1WC2g86vqA1UvdKJVNCU9td1TR8j8wQXo",
+	"3tf2MkQzmLHBjnsmRwKnJQgASrmUggxK2Q3PiFd99tDn",
+	"3X31YYsRw8We2YhsK29QtVwXXk783HbbYDGAV1HBjcBD",
+	"FhDjU5r4MWx6KdfY7MVx6w1YJf6tvRJyj6mbxqU8N3F7",
+	"57xqRiBeQjHgrYqhXUHRKh9WU9Ukya79U3hM7QPr13Gy",
+	"7pjSLC42Er4KPdVLZW7VGkxU9tLZKBqt4apwyZzuWyYU",
+	"jyDrCsnuvxGM9H7YE2rLsBYoFoSHWABFoi7P61FaQ3Q",
+	"GUHXNHugMc22rkX6Mz4GMU4Vj1hbPa3DCrfCDRQFWQ2b",
+	"2azGMpfp91pqd5gXpZJWK8egdpgUxDkhXK8UYHtRjiZa",
+	"CPH4QdmL4yNB9KBpr3bQwUQxdQbMeUKsUyos2ViHaNTB",
+	"GKNqfGFsK1Th32GSRkT9kaiA7w89GKJJmGVV5ibo8xn2",
+	"GzuWB5nf2NH15Ssk15n3Zd72iykMoPm8Qx5TPCUS99LC",
+	"7eAojuq3vcrev41DuVFgpZ1yagQhpUNhUn2uXnKo7A41",
+	"BCuFo9AhTREJ5bgJzCrhXzckmAoyscrDDkggixX6t3c5",
+	"FUnpGc7v43bvBvC584gQXiRdcuCMnDoXLXbJmMNkg3wQ",
+	"FtFPeHGXZhgacdNoXh2dYKBZkgmTq9YYoUZX97hyjgh4",
+	"2f3V4h5z9jds59EeFVqViVKuZrMoYM3xb3eq8fWWuN7Y",
+	"BEvgtgRX7DdUrZ8Jrw5SMLctA7pQ76ScGry73mEzH869",
+	"FD8pHBAwhq2VtHQQSTpdpnmiEoMNfFzJhQXGhSQVkdcQ",
+	"HqQuQ9wF3QE7RwiYdgB88SnHwi1n5Q2ogidy7WfZJGgb",
+	"Dg2fDYcuvRxCBtZtf1rB2bgbc4KTmy8KMC22Y4JFX7Qd",
+	"p6pGSE2rLDH7yiiZ7bKoNZcB3YaszRsRDQ5rVRuTXiz",
+	"HEg2w4Ev5ouoZB51Tmhj4DBPG7jrxKaTrf9GfKYubBbG",
+	"A5mcmJHSMARvaQcYXGQ96Nx1h4sFeReJNTCBxqxxMqrF",
+}
+
+// ConnectedPdaInfo is struct representing example connected programs pda
+type ConnectedPdaInfo struct {
+	Discriminator     [8]byte
+	LastSender        ethcommon.Address
+	LastMessage       string
+	LastRevertSender  solana.PublicKey
+	LastRevertMessage string
+}
+
+// ParseConnectedPda deserializes connectedPda into ConnectedPdaInfo struct
+func (r *E2ERunner) ParseConnectedPda(connectedPda solana.PublicKey) ConnectedPdaInfo {
+	connectedPdaInfo, err := r.SolanaClient.GetAccountInfo(r.Ctx, connectedPda)
+	require.NoError(r, err)
+	pda := ConnectedPdaInfo{}
+	err = borsh.Deserialize(&pda, connectedPdaInfo.Bytes())
+	require.NoError(r, err)
+
+	return pda
+}
+
 // ComputePdaAddress computes the PDA address for the gateway program
 func (r *E2ERunner) ComputePdaAddress() solana.PublicKey {
 	seed := []byte(solanacontract.PDASeed)
@@ -791,127 +872,9 @@ func (r *E2ERunner) WaitAndVerifySPLBalanceChange(
 	}
 }
 
-// SetupTestALTWithRandomWallets sets up ALT with random accounts provided in setup solana script, with accounts provided as argument
-// used to test ALTs with large amount of accounts
-func (r *E2ERunner) SetupTestALTWithRandomWallets(accounts []solana.PublicKey) (solana.PublicKey, []solana.PublicKey) {
-	privkey := r.GetSolanaPrivKey()
-
-	recentSlot, err := r.SolanaClient.GetSlot(r.Ctx, rpc.CommitmentFinalized)
-	require.NoError(r, err)
-
-	// prefunded random wallets used for testing big number of accounts in payload
-	randomWallets := []solana.PublicKey{
-		solana.MustPublicKeyFromBase58("4C2kkMnqXMfPJ8PPK5v6TCg42k5z16f2kzqVocXoSDcq"),
-		solana.MustPublicKeyFromBase58("3tqcVCVz5Jztwnku1H9zpvjaWshSpHakMuX4xUJQuuuA"),
-		solana.MustPublicKeyFromBase58("7duGsuv6nB3yr15EuWuHEDD7rWovpAnjuveXJ5ySZuFV"),
-		solana.MustPublicKeyFromBase58("8vjuCrCKVfnBGinWjc33zLRnG8iy53wj3YWHhKqvTE7o"),
-		solana.MustPublicKeyFromBase58("bzkoxG5YMeWxKfNcjzbEHb3XaGTY4NLKfejjDmxVhhY"),
-		solana.MustPublicKeyFromBase58("GUjKWPmnXNwPLR6kcrkLSBARmQdYnySPpxVNEUGFLs72"),
-		solana.MustPublicKeyFromBase58("5oqdTyA78hpeP8RTwRBmoxCvp1V7DFicKj7T2DvtDDQM"),
-		solana.MustPublicKeyFromBase58("C481t79gpWbsWwPD9eJZTAo5TSaTBet8icEkiPhwKLDx"),
-		solana.MustPublicKeyFromBase58("EJvNNovWkfQYrmyMncqVvHNde2QJpSA4EJk355vyQWph"),
-		solana.MustPublicKeyFromBase58("EkUpd7HFbSYPJEDbXZeDsCG19Hj5vbKTUt4rzpPYKsTM"),
-		solana.MustPublicKeyFromBase58("7c7TqqdbKRWDNLVAxNa481F355AAij1fdSRttzUNVNeD"),
-		solana.MustPublicKeyFromBase58("FuefjNTywey57U2zW6SmBWaJsCx7E84jmWUCbk52sBHR"),
-		solana.MustPublicKeyFromBase58("fczqc5N5arnKbvMj1kgg9P1FpYPQBmJsyTXcrmK9bbp"),
-		solana.MustPublicKeyFromBase58("GyH4mpobR6g2npNo5vRNcs2Cv8CxDKbK82p34kNCB4p2"),
-		solana.MustPublicKeyFromBase58("9XYF8U1srAkUETkywtFrZsApipENiMU3C8Rnz5aPib94"),
-		solana.MustPublicKeyFromBase58("ABnm3PMB4onvCFriWg7eBcNSmiye9iq9rRdeBViJqWif"),
-		solana.MustPublicKeyFromBase58("AEydLk3RXZv67wry7EMZmS1uHLYdH8ia6xdsYri4hyB2"),
-		solana.MustPublicKeyFromBase58("ArDNFdmDzrRP13UTJ4nyP11NfjV6aiQrxNwFUnwm3h8N"),
-		solana.MustPublicKeyFromBase58("FzAt9aPKFUy1D2Qq8u7myYW8HFqKLzbQS2paaWz8iAmZ"),
-		solana.MustPublicKeyFromBase58("9RybduN4CJHaZXvUiHoZ7KsHS9dgv1NCSAdJZaRJDW5U"),
-		solana.MustPublicKeyFromBase58("C54jMgtk2umaJYoPD8aF3hmuH8XkAz2xA4sxr2ZtJABV"),
-		solana.MustPublicKeyFromBase58("8kRqLbezvj4apyaK6fanurQJjhDQwn6wnW1Yr9H96gsT"),
-		solana.MustPublicKeyFromBase58("2fddFSJoGJ2YuAZWxEvK9pRXXWLSKJ44rfJNVZg5WHBn"),
-		solana.MustPublicKeyFromBase58("GW9oi4yqAUFBcUNUHqz56FRdfHU6Md1t9o7i2svL1XcG"),
-		solana.MustPublicKeyFromBase58("Gre1nqrE1KyBbBH2Xb4qVQWFFdyXetLR5rruJpJkNhkV"),
-		solana.MustPublicKeyFromBase58("BUrJRTsFVqnuLeq4Vfs7NrrU6EsL957n4tBZGoRRxkza"),
-		solana.MustPublicKeyFromBase58("3tE2kKyPfuwC5rZUpJ9NaKMBf5og7G3rVTzE9CNMm9Q9"),
-		solana.MustPublicKeyFromBase58("CFEQ79VSAupXWmdzvmjNtef5BybDVYXBKeh8frNHZiYe"),
-		solana.MustPublicKeyFromBase58("bTnsajQuybXV6Wf9V7a8wQrwaWZ1WskkmUUChimkWmc"),
-		solana.MustPublicKeyFromBase58("4dpxhhomWY3A9ey9g7EfxRQKfXnikfM4tRrFAgQf8Y8n"),
-		solana.MustPublicKeyFromBase58("aQJPrcj4LNNHh9UK41sfcACFspaFR7wgcUTgmSKRXiB"),
-		solana.MustPublicKeyFromBase58("5ndhaFZ48eKyU7f66vq7WSbjZRh9WhpnbgdMwDRrvgj4"),
-		solana.MustPublicKeyFromBase58("G7m7dSWH5tb1WC2g86vqA1UvdKJVNCU9td1TR8j8wQXo"),
-		solana.MustPublicKeyFromBase58("3tf2MkQzmLHBjnsmRwKnJQgASrmUggxK2Q3PiFd99tDn"),
-		solana.MustPublicKeyFromBase58("3X31YYsRw8We2YhsK29QtVwXXk783HbbYDGAV1HBjcBD"),
-		solana.MustPublicKeyFromBase58("FhDjU5r4MWx6KdfY7MVx6w1YJf6tvRJyj6mbxqU8N3F7"),
-		solana.MustPublicKeyFromBase58("57xqRiBeQjHgrYqhXUHRKh9WU9Ukya79U3hM7QPr13Gy"),
-		solana.MustPublicKeyFromBase58("7pjSLC42Er4KPdVLZW7VGkxU9tLZKBqt4apwyZzuWyYU"),
-		solana.MustPublicKeyFromBase58("jyDrCsnuvxGM9H7YE2rLsBYoFoSHWABFoi7P61FaQ3Q"),
-		solana.MustPublicKeyFromBase58("GUHXNHugMc22rkX6Mz4GMU4Vj1hbPa3DCrfCDRQFWQ2b"),
-		solana.MustPublicKeyFromBase58("2azGMpfp91pqd5gXpZJWK8egdpgUxDkhXK8UYHtRjiZa"),
-		solana.MustPublicKeyFromBase58("CPH4QdmL4yNB9KBpr3bQwUQxdQbMeUKsUyos2ViHaNTB"),
-		solana.MustPublicKeyFromBase58("GKNqfGFsK1Th32GSRkT9kaiA7w89GKJJmGVV5ibo8xn2"),
-		solana.MustPublicKeyFromBase58("GzuWB5nf2NH15Ssk15n3Zd72iykMoPm8Qx5TPCUS99LC"),
-		solana.MustPublicKeyFromBase58("7eAojuq3vcrev41DuVFgpZ1yagQhpUNhUn2uXnKo7A41"),
-		solana.MustPublicKeyFromBase58("BCuFo9AhTREJ5bgJzCrhXzckmAoyscrDDkggixX6t3c5"),
-		solana.MustPublicKeyFromBase58("FUnpGc7v43bvBvC584gQXiRdcuCMnDoXLXbJmMNkg3wQ"),
-		solana.MustPublicKeyFromBase58("FtFPeHGXZhgacdNoXh2dYKBZkgmTq9YYoUZX97hyjgh4"),
-		solana.MustPublicKeyFromBase58("2f3V4h5z9jds59EeFVqViVKuZrMoYM3xb3eq8fWWuN7Y"),
-		solana.MustPublicKeyFromBase58("BEvgtgRX7DdUrZ8Jrw5SMLctA7pQ76ScGry73mEzH869"),
-		solana.MustPublicKeyFromBase58("FD8pHBAwhq2VtHQQSTpdpnmiEoMNfFzJhQXGhSQVkdcQ"),
-		solana.MustPublicKeyFromBase58("HqQuQ9wF3QE7RwiYdgB88SnHwi1n5Q2ogidy7WfZJGgb"),
-		solana.MustPublicKeyFromBase58("Dg2fDYcuvRxCBtZtf1rB2bgbc4KTmy8KMC22Y4JFX7Qd"),
-		solana.MustPublicKeyFromBase58("p6pGSE2rLDH7yiiZ7bKoNZcB3YaszRsRDQ5rVRuTXiz"),
-		solana.MustPublicKeyFromBase58("HEg2w4Ev5ouoZB51Tmhj4DBPG7jrxKaTrf9GfKYubBbG"),
-		solana.MustPublicKeyFromBase58("A5mcmJHSMARvaQcYXGQ96Nx1h4sFeReJNTCBxqxxMqrF"),
-	}
-
-	accounts = append(accounts, randomWallets...)
-
-	buf := make([]byte, 8)
-	binary.LittleEndian.PutUint64(buf, recentSlot)
-
-	altAddress, bump, err := solana.FindProgramAddress(
-		[][]byte{privkey.PublicKey().Bytes(), buf},
-		solana.AddressLookupTableProgramID,
-	)
-	require.NoError(r, err)
-
-	// create ALT and extend 2 times due to large amount of accounts
-	createALTInstruction := addresslookuptable.NewCreateAddressLookupTableInstruction(recentSlot, bump, altAddress, privkey.PublicKey(), privkey.PublicKey()).
-		Build()
-
-	signedTx := r.CreateSignedTransaction(
-		[]solana.Instruction{createALTInstruction},
-		privkey,
-		[]solana.PrivateKey{},
-	)
-	r.BroadcastTxSync(signedTx)
-
-	// need to wait a bit for ALT to be active
-	time.Sleep(1 * time.Second)
-
-	extendALTInstruction := addresslookuptable.NewExtendAddressLookupTableInstruction(accounts[:30], altAddress, privkey.PublicKey(), privkey.PublicKey()).
-		Build()
-
-	signedTx = r.CreateSignedTransaction(
-		[]solana.Instruction{extendALTInstruction},
-		privkey,
-		[]solana.PrivateKey{},
-	)
-	r.BroadcastTxSync(signedTx)
-
-	time.Sleep(1 * time.Second)
-
-	extendALTInstruction = addresslookuptable.NewExtendAddressLookupTableInstruction(accounts[30:], altAddress, privkey.PublicKey(), privkey.PublicKey()).
-		Build()
-
-	signedTx = r.CreateSignedTransaction(
-		[]solana.Instruction{extendALTInstruction},
-		privkey,
-		[]solana.PrivateKey{},
-	)
-	r.BroadcastTxSync(signedTx)
-
-	return altAddress, randomWallets
-}
-
-// SetupTestALTWithRandomWallets sets up ALT with random accounts provided in setup solana script, with accounts provided as argument
-// using only 10 accounts because this is creating ATAs, and point is not to test accounts amount, just execute SPL functionality
-func (r *E2ERunner) SetupTestALTWithRandomWalletsSPL(
+// SetupTestAddressLookupTableWithRandomWallets sets up AddressLookupTable with random accounts provided in setup solana script, with accounts provided as argument
+// used to test AddressLookupTables with large amount of accounts
+func (r *E2ERunner) SetupTestAddressLookupTableWithRandomWallets(
 	accounts []solana.PublicKey,
 ) (solana.PublicKey, []solana.PublicKey) {
 	privkey := r.GetSolanaPrivKey()
@@ -920,17 +883,77 @@ func (r *E2ERunner) SetupTestALTWithRandomWalletsSPL(
 	require.NoError(r, err)
 
 	// prefunded random wallets used for testing big number of accounts in payload
-	randomWallets := []solana.PublicKey{
-		solana.MustPublicKeyFromBase58("4C2kkMnqXMfPJ8PPK5v6TCg42k5z16f2kzqVocXoSDcq"),
-		solana.MustPublicKeyFromBase58("3tqcVCVz5Jztwnku1H9zpvjaWshSpHakMuX4xUJQuuuA"),
-		solana.MustPublicKeyFromBase58("7duGsuv6nB3yr15EuWuHEDD7rWovpAnjuveXJ5ySZuFV"),
-		solana.MustPublicKeyFromBase58("8vjuCrCKVfnBGinWjc33zLRnG8iy53wj3YWHhKqvTE7o"),
-		solana.MustPublicKeyFromBase58("bzkoxG5YMeWxKfNcjzbEHb3XaGTY4NLKfejjDmxVhhY"),
-		solana.MustPublicKeyFromBase58("GUjKWPmnXNwPLR6kcrkLSBARmQdYnySPpxVNEUGFLs72"),
-		solana.MustPublicKeyFromBase58("5oqdTyA78hpeP8RTwRBmoxCvp1V7DFicKj7T2DvtDDQM"),
-		solana.MustPublicKeyFromBase58("C481t79gpWbsWwPD9eJZTAo5TSaTBet8icEkiPhwKLDx"),
-		solana.MustPublicKeyFromBase58("EJvNNovWkfQYrmyMncqVvHNde2QJpSA4EJk355vyQWph"),
-		solana.MustPublicKeyFromBase58("EkUpd7HFbSYPJEDbXZeDsCG19Hj5vbKTUt4rzpPYKsTM"),
+	// they are randomly generated and funded in start-solana.sh script
+	randomWallets := make([]solana.PublicKey, len(prefundedRandomWalletAddresses))
+	for i, addr := range prefundedRandomWalletAddresses {
+		randomWallets[i] = solana.MustPublicKeyFromBase58(addr)
+	}
+
+	accounts = append(accounts, randomWallets...)
+
+	buf := make([]byte, 8)
+	binary.LittleEndian.PutUint64(buf, recentSlot)
+
+	addressLookupTableAddress, bump, err := solana.FindProgramAddress(
+		[][]byte{privkey.PublicKey().Bytes(), buf},
+		solana.AddressLookupTableProgramID,
+	)
+	require.NoError(r, err)
+
+	// create AddressLookupTable and extend 2 times due to large amount of accounts
+	createAddressLookupTableInstruction := addresslookuptable.NewCreateAddressLookupTableInstruction(recentSlot, bump, addressLookupTableAddress, privkey.PublicKey(), privkey.PublicKey()).
+		Build()
+
+	signedTx := r.CreateSignedTransaction(
+		[]solana.Instruction{createAddressLookupTableInstruction},
+		privkey,
+		[]solana.PrivateKey{},
+	)
+	r.BroadcastTxSync(signedTx)
+
+	// need to wait a bit for AddressLookupTable to be active
+	time.Sleep(1 * time.Second)
+
+	extendAddressLookupTableInstruction := addresslookuptable.NewExtendAddressLookupTableInstruction(accounts[:30], addressLookupTableAddress, privkey.PublicKey(), privkey.PublicKey()).
+		Build()
+
+	signedTx = r.CreateSignedTransaction(
+		[]solana.Instruction{extendAddressLookupTableInstruction},
+		privkey,
+		[]solana.PrivateKey{},
+	)
+	r.BroadcastTxSync(signedTx)
+
+	time.Sleep(1 * time.Second)
+
+	extendAddressLookupTableInstruction = addresslookuptable.NewExtendAddressLookupTableInstruction(accounts[30:], addressLookupTableAddress, privkey.PublicKey(), privkey.PublicKey()).
+		Build()
+
+	signedTx = r.CreateSignedTransaction(
+		[]solana.Instruction{extendAddressLookupTableInstruction},
+		privkey,
+		[]solana.PrivateKey{},
+	)
+	r.BroadcastTxSync(signedTx)
+
+	return addressLookupTableAddress, randomWallets
+}
+
+// SetupTestAddressLookupTableWithRandomWallets sets up AddressLookupTable with random accounts provided in setup solana script, with accounts provided as argument
+// using only 10 accounts because this is creating ATAs, and point is not to test accounts amount, just execute SPL functionality
+func (r *E2ERunner) SetupTestAddressLookupTableWithRandomWalletsSPL(
+	accounts []solana.PublicKey,
+) (solana.PublicKey, []solana.PublicKey) {
+	privkey := r.GetSolanaPrivKey()
+
+	recentSlot, err := r.SolanaClient.GetSlot(r.Ctx, rpc.CommitmentFinalized)
+	require.NoError(r, err)
+
+	// prefunded random wallets used for testing big number of accounts in payload
+	// they are randomly generated and funded in start-solana.sh script
+	randomWallets := make([]solana.PublicKey, 10)
+	for i := 0; i < 10; i++ {
+		randomWallets[i] = solana.MustPublicKeyFromBase58(prefundedRandomWalletAddresses[i])
 	}
 
 	randomWalletsAta := []solana.PublicKey{}
@@ -944,31 +967,31 @@ func (r *E2ERunner) SetupTestALTWithRandomWalletsSPL(
 	buf := make([]byte, 8)
 	binary.LittleEndian.PutUint64(buf, recentSlot)
 
-	altAddress, bump, err := solana.FindProgramAddress(
+	addressLookupTableAddress, bump, err := solana.FindProgramAddress(
 		[][]byte{privkey.PublicKey().Bytes(), buf},
 		solana.AddressLookupTableProgramID,
 	)
 	require.NoError(r, err)
 
-	// create ALT and extend it
-	createALTInstruction := addresslookuptable.NewCreateAddressLookupTableInstruction(recentSlot, bump, altAddress, privkey.PublicKey(), privkey.PublicKey()).
+	// create AddressLookupTable and extend it
+	createAddressLookupTableInstruction := addresslookuptable.NewCreateAddressLookupTableInstruction(recentSlot, bump, addressLookupTableAddress, privkey.PublicKey(), privkey.PublicKey()).
 		Build()
 
 	signedTx := r.CreateSignedTransaction(
-		[]solana.Instruction{createALTInstruction},
+		[]solana.Instruction{createAddressLookupTableInstruction},
 		privkey,
 		[]solana.PrivateKey{},
 	)
 	r.BroadcastTxSync(signedTx)
 
-	// need to wait a bit for ALT to be active
+	// need to wait a bit for AddressLookupTable to be active
 	time.Sleep(1 * time.Second)
 
-	extendALTInstruction := addresslookuptable.NewExtendAddressLookupTableInstruction(accounts, altAddress, privkey.PublicKey(), privkey.PublicKey()).
+	extendAddressLookupTableInstruction := addresslookuptable.NewExtendAddressLookupTableInstruction(accounts, addressLookupTableAddress, privkey.PublicKey(), privkey.PublicKey()).
 		Build()
 
 	signedTx = r.CreateSignedTransaction(
-		[]solana.Instruction{extendALTInstruction},
+		[]solana.Instruction{extendAddressLookupTableInstruction},
 		privkey,
 		[]solana.PrivateKey{},
 	)
@@ -976,5 +999,5 @@ func (r *E2ERunner) SetupTestALTWithRandomWalletsSPL(
 
 	time.Sleep(1 * time.Second)
 
-	return altAddress, randomWalletsAta
+	return addressLookupTableAddress, randomWalletsAta
 }