+++ pkg/contracts/solana/gateway_message.go
@@ -266,11 +266,11 @@ type MsgExecute struct {
 	// Remaining accounts for arbirtrary program
 	remainingAccounts []*solana.AccountMeta
 
-	// Address of ALT
-	altAddress *solana.PublicKey
+	// Address of AddressLookupTable
+	addressLookupTableAddress *solana.PublicKey
 
-	// Addresses in ALT state
-	altStateAddresses solana.PublicKeySlice
+	// Addresses in AddressLookupTable state
+	addressLookupTableStateAddresses solana.PublicKeySlice
 }
 
 // NewMsgExecute returns a new execute message
@@ -281,20 +281,20 @@ func NewMsgExecute(
 	data []byte,
 	executeType ExecuteType,
 	remainingAccounts []*solana.AccountMeta,
-	altAddress *solana.PublicKey,
-	altStateAddresses solana.PublicKeySlice,
+	addressLookupTableAddress *solana.PublicKey,
+	addressLookupTableStateAddresses solana.PublicKeySlice,
 ) *MsgExecute {
 	return &MsgExecute{
-		chainID:           chainID,
-		nonce:             nonce,
-		amount:            amount,
-		to:                to,
-		sender:            sender,
-		data:              data,
-		executeType:       executeType,
-		remainingAccounts: remainingAccounts,
-		altAddress:        altAddress,
-		altStateAddresses: altStateAddresses,
+		chainID:                          chainID,
+		nonce:                            nonce,
+		amount:                           amount,
+		to:                               to,
+		sender:                           sender,
+		data:                             data,
+		executeType:                      executeType,
+		remainingAccounts:                remainingAccounts,
+		addressLookupTableAddress:        addressLookupTableAddress,
+		addressLookupTableStateAddresses: addressLookupTableStateAddresses,
 	}
 }
 
@@ -338,14 +338,14 @@ func (msg *MsgExecute) ExecuteType() ExecuteType {
 	return msg.executeType
 }
 
-// ALT return address of ALT
-func (msg *MsgExecute) ALT() *solana.PublicKey {
-	return msg.altAddress
+// AddressLookupTable return address of AddressLookupTable
+func (msg *MsgExecute) AddressLookupTable() *solana.PublicKey {
+	return msg.addressLookupTableAddress
 }
 
-// ALTStateAddresses returns addresses from ALT state
-func (msg *MsgExecute) ALTStateAddresses() solana.PublicKeySlice {
-	return msg.altStateAddresses
+// AddressLookupTableStateAddresses returns addresses from AddressLookupTable state
+func (msg *MsgExecute) AddressLookupTableStateAddresses() solana.PublicKeySlice {
+	return msg.addressLookupTableStateAddresses
 }
 
 // Hash packs the execute message and computes the hash
@@ -583,11 +583,11 @@ type MsgExecuteSPL struct {
 	// Remaining accounts for arbirtrary program
 	remainingAccounts []*solana.AccountMeta
 
-	// Address of ALT
-	altAddress *solana.PublicKey
+	// Address of AddressLookupTable
+	addressLookupTableAddress *solana.PublicKey
 
-	// Addresses in ALT state
-	altStateAddresses solana.PublicKeySlice
+	// Addresses in AddressLookupTable state
+	addressLookupTableStateAddresses solana.PublicKeySlice
 }
 
 // NewMsgExecuteSPL returns a new execute spl message
@@ -599,23 +599,23 @@ func NewMsgExecuteSPL(
 	data []byte,
 	executeType ExecuteType,
 	remainingAccounts []*solana.AccountMeta,
-	altAddress *solana.PublicKey,
-	altStateAddresses solana.PublicKeySlice,
+	addressLookupTableAddress *solana.PublicKey,
+	addressLookupTableStateAddresses solana.PublicKeySlice,
 ) *MsgExecuteSPL {
 	return &MsgExecuteSPL{
-		chainID:           chainID,
-		nonce:             nonce,
-		amount:            amount,
-		to:                to,
-		recipientAta:      toAta,
-		mintAccount:       mintAccount,
-		decimals:          decimals,
-		sender:            sender,
-		data:              data,
-		executeType:       executeType,
-		remainingAccounts: remainingAccounts,
-		altAddress:        altAddress,
-		altStateAddresses: altStateAddresses,
+		chainID:                          chainID,
+		nonce:                            nonce,
+		amount:                           amount,
+		to:                               to,
+		recipientAta:                     toAta,
+		mintAccount:                      mintAccount,
+		decimals:                         decimals,
+		sender:                           sender,
+		data:                             data,
+		executeType:                      executeType,
+		remainingAccounts:                remainingAccounts,
+		addressLookupTableAddress:        addressLookupTableAddress,
+		addressLookupTableStateAddresses: addressLookupTableStateAddresses,
 	}
 }
 
@@ -666,14 +666,14 @@ func (msg *MsgExecuteSPL) RemainingAccounts() []*solana.AccountMeta {
 	return msg.remainingAccounts
 }
 
-// ALT return address of ALT
-func (msg *MsgExecuteSPL) ALT() *solana.PublicKey {
-	return msg.altAddress
+// AddressLookupTable return address of AddressLookupTable
+func (msg *MsgExecuteSPL) AddressLookupTable() *solana.PublicKey {
+	return msg.addressLookupTableAddress
 }
 
-// ALTStateAddresses returns addresses from ALT state
-func (msg *MsgExecuteSPL) ALTStateAddresses() solana.PublicKeySlice {
-	return msg.altStateAddresses
+// AddressLookupTableStateAddresses returns addresses from AddressLookupTable state
+func (msg *MsgExecuteSPL) AddressLookupTableStateAddresses() solana.PublicKeySlice {
+	return msg.addressLookupTableStateAddresses
 }
 
 // ExecuteType returns the type of execute operation