+++ changelog.md
@@ -20,7 +20,7 @@ by calling `updateAdditionalActionFee` admin function.
 ### Features
 
 * [4274](https://github.com/zeta-chain/node/pull/4274) - multiple evm calls in single tx
-* [4266](https://github.com/zeta-chain/node/pull/4266) - add solana ALT support 
+* [4266](https://github.com/zeta-chain/node/pull/4266) - add Solana Address Lookup Table support 
 * [4288](https://github.com/zeta-chain/node/pull/4288) - zetaclient config feature flag for multiple evm calls
 
 ## v36.0.0